019 Authorization with Gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\AttendeeResource;
 use App\Models\Event;
 use Illuminate\Http\Request;
 use App\Http\Traits\CanLoadRelationships;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -61,8 +62,12 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        if (Gate::denies('delete-attendee', [$event, $attendee])) {
+            abort(403, 'You are not authorized to update this event!');
+        };
+
         $attendee->delete();
 
         return response(status: 204);
